Add files via upload
I still have some cleaning up to do, but here is the updated version of the fuzzy rules. There are 2 types some that are hand created, and a version that is automatically generated  into low middle and high. That might be better for some variables, and it might give us a point of comparison for comparing the population we have to what is recommended. I will work on cleaning this up this weekend and doing some clustering. If you need the data in a particular format, please let me know. I mainly worked with pandas dataframes, but in this current iteration, the handcrafted membership functions are numpy arrays, and the automatically generated ones are lists within dictionaries. I should be able to change formats no problem. Let me know if you have any questions.